requirements.txt
fastapi
uvicorn[standard]
pywebpush
cryptography
pillow
requests
